requirements.txt
mysql-connector-python=2.1.5